backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
firebase-admin>=6.0.0
vnstock>=3.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
APScheduler>=3.10.0
